make all hashes uint64_t with correct fnv-1a parameters

+++ strings.c
@@ -56,11 +56,11 @@ void Str_debug(const Str *str) {
   printf("%.*s", (int)str->len, str->p);
 }
 
-inline uint32_t Str_hash(const Str *str) {
+inline uint64_t Str_hash(const Str *str) {
   // https://en.wikipedia.org/wiki/Fowler%E2%80%93Noll%E2%80%93Vo_hash_function#FNV-1a_hash
   // https://en.wikipedia.org/wiki/Fowler%E2%80%93Noll%E2%80%93Vo_hash_function#FNV_hash_parameters
-  uint32_t hash = FNV_OFFSET_BASIS;
-  for (uint32_t i = 0; i < str->len; i++) {
+  uint64_t hash = FNV_OFFSET_BASIS;
+  for (size_t i = 0; i < str->len; i++) {
     hash ^= str->p[i];
     hash *= FNV_PRIME;
   }
